+++ src/InvestmentApp.Application/MarketData/Queries/GetTradingHistorySummary/GetTradingHistorySummaryQuery.cs
@@ -11,9 +11,9 @@ public class GetTradingHistorySummaryQuery : IRequest<TradingHistorySummaryDto>
 public class TradingHistorySummaryDto
 {
     public string Symbol { get; set; } = null!;
-    public decimal ChangeDay { get; set; }
-    public decimal ChangeWeek { get; set; }
-    public decimal ChangeMonth { get; set; }
+    public decimal? ChangeDay { get; set; }
+    public decimal? ChangeWeek { get; set; }
+    public decimal? ChangeMonth { get; set; }
     public decimal? Change3Month { get; set; }
     public decimal? Change6Month { get; set; }
 }